Remove DropWrapper
The reason this was a standalone component is that drop handling is
likely to be very custom and I didn't want to bake in our own use case.
But the recipe is simple enough I don't think the maintenance outweighs
the convenience; I'd rather just provide a how-to in the wiki or
something.